refactor!: split ajax request and handling into separate classes
