+++ src/BackgroundProcess/Domain/BackgroundJobQueue.php
@@ -1,31 +1,31 @@
 <?php
 declare(strict_types=1);
 
-namespace Jetty\BackgroundProcessing\BackgroundProcess\Application\Port\Out;
+namespace Jetty\BackgroundProcessing\BackgroundProcess\Domain;
 
 use Jetty\BackgroundProcessing\BackgroundProcess\Domain\AsyncRequest;
 
 /**
  * Describes the minimum functionality needed for an async request.
  */
-interface BackgroundProcess extends AsyncRequest
+interface BackgroundJobQueue extends AsyncRequest
 {
     /**
      * Push an array of job data to the queue.
      *
      * @param array $data Data.
      */
-    public function push_to_queue(array $data): BackgroundProcess;
+    public function pushToQueue(array $data): BackgroundJobQueue;
 
 
     /**
      * Save the queue to be processed later.
      */
-    public function save(): BackgroundProcess;
+    public function save(): BackgroundJobQueue;
 
 
     /**
      * Cancel the background process.
      */
-    public function cancel_process(): void;
+    public function cancel(): void;
 }
